functions/gcf-download-file/requirements.txt
google-cloud-storage>=1.6.0
google-cloud-datastore>=1.6.0
python-slugify